+++ framework-test/src/main/java/org/checkerframework/framework/test/TypecheckExecutor.java
@@ -2,7 +2,6 @@ package org.checkerframework.framework.test;
 
 import org.checkerframework.framework.test.diagnostics.JavaDiagnosticReader;
 import org.checkerframework.framework.test.diagnostics.TestDiagnostic;
-import org.checkerframework.javacutil.SystemUtil;
 import org.plumelib.util.StringsPlume;
 
 import java.io.File;
@@ -75,12 +74,6 @@ public class TypecheckExecutor {
             final List<String> options = new ArrayList<>();
             options.add("-processor");
             options.add(String.join(",", configuration.getProcessors()));
-            if (SystemUtil.jreVersion == 8) {
-                options.add("-source");
-                options.add("8");
-                options.add("-target");
-                options.add("8");
-            }
 
             List<String> nonJvmOptions = new ArrayList<>();
             for (String option : configuration.getFlatOptions()) {