+++ framework/src/main/java/org/checkerframework/common/wholeprograminference/scenelib/ASceneWrapper.java
@@ -2,14 +2,14 @@ package org.checkerframework.common.wholeprograminference.scenelib;
 
 import com.sun.tools.javac.code.Symbol.ClassSymbol;
 
-import org.checkerframework.afu.scenelib.annotations.Annotation;
-import org.checkerframework.afu.scenelib.annotations.el.AClass;
-import org.checkerframework.afu.scenelib.annotations.el.AField;
-import org.checkerframework.afu.scenelib.annotations.el.AMethod;
-import org.checkerframework.afu.scenelib.annotations.el.AScene;
-import org.checkerframework.afu.scenelib.annotations.el.ATypeElement;
-import org.checkerframework.afu.scenelib.annotations.el.DefException;
-import org.checkerframework.afu.scenelib.annotations.io.IndexFileWriter;
+import org.checkerframework.afu.scenelib.Annotation;
+import org.checkerframework.afu.scenelib.el.AClass;
+import org.checkerframework.afu.scenelib.el.AField;
+import org.checkerframework.afu.scenelib.el.AMethod;
+import org.checkerframework.afu.scenelib.el.AScene;
+import org.checkerframework.afu.scenelib.el.ATypeElement;
+import org.checkerframework.afu.scenelib.el.DefException;
+import org.checkerframework.afu.scenelib.io.IndexFileWriter;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.common.basetype.BaseTypeChecker;
 import org.checkerframework.common.wholeprograminference.AnnotationConverter;
@@ -27,7 +27,6 @@ import org.plumelib.util.CollectionsPlume;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -203,12 +202,7 @@ public class ASceneWrapper {
             return;
         }
         if (classSymbol.isEnum()) {
-            List<VariableElement> enumConstants = new ArrayList<>();
-            for (Element e : classSymbol.getEnclosedElements()) {
-                if (e.getKind() == ElementKind.ENUM_CONSTANT) {
-                    enumConstants.add((VariableElement) e);
-                }
-            }
+            List<VariableElement> enumConstants = ElementUtils.getEnumConstants(classSymbol);
             if (!aClass.isEnum(classSymbol.getSimpleName().toString())) {
                 aClass.setEnumConstants(enumConstants);
             } else {