+++ framework/src/test/java/org/checkerframework/framework/testchecker/lubglb/LubGlbAnnotatedTypeFactory.java
@@ -2,13 +2,13 @@ package org.checkerframework.framework.testchecker.lubglb;
 
 import org.checkerframework.common.basetype.BaseAnnotatedTypeFactory;
 import org.checkerframework.common.basetype.BaseTypeChecker;
-import org.checkerframework.framework.testchecker.lubglb.quals.A;
-import org.checkerframework.framework.testchecker.lubglb.quals.B;
-import org.checkerframework.framework.testchecker.lubglb.quals.C;
-import org.checkerframework.framework.testchecker.lubglb.quals.D;
-import org.checkerframework.framework.testchecker.lubglb.quals.E;
-import org.checkerframework.framework.testchecker.lubglb.quals.F;
-import org.checkerframework.framework.testchecker.lubglb.quals.Poly;
+import org.checkerframework.framework.testchecker.lubglb.quals.LubglbA;
+import org.checkerframework.framework.testchecker.lubglb.quals.LubglbB;
+import org.checkerframework.framework.testchecker.lubglb.quals.LubglbC;
+import org.checkerframework.framework.testchecker.lubglb.quals.LubglbD;
+import org.checkerframework.framework.testchecker.lubglb.quals.LubglbE;
+import org.checkerframework.framework.testchecker.lubglb.quals.LubglbF;
+import org.checkerframework.framework.testchecker.lubglb.quals.PolyLubglb;
 
 import java.lang.annotation.Annotation;
 import java.util.Arrays;
@@ -25,6 +25,13 @@ public class LubGlbAnnotatedTypeFactory extends BaseAnnotatedTypeFactory {
     @Override
     protected Set<Class<? extends Annotation>> createSupportedTypeQualifiers() {
         return new HashSet<Class<? extends Annotation>>(
-                Arrays.asList(A.class, B.class, C.class, D.class, E.class, F.class, Poly.class));
+                Arrays.asList(
+                        LubglbA.class,
+                        LubglbB.class,
+                        LubglbC.class,
+                        LubglbD.class,
+                        LubglbE.class,
+                        LubglbF.class,
+                        PolyLubglb.class));
     }
 }