+++ framework/src/test/java/org/checkerframework/framework/testchecker/reflection/qual/TestReflectBottom.java
@@ -1,0 +1,21 @@
+package org.checkerframework.framework.testchecker.reflection.qual;
+
+import org.checkerframework.framework.qual.DefaultFor;
+import org.checkerframework.framework.qual.SubtypeOf;
+import org.checkerframework.framework.qual.TargetLocations;
+import org.checkerframework.framework.qual.TypeUseLocation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Target;
+
+/**
+ * Toy type system for testing reflection resolution. Uses
+ * org.checkerframework.common.subtyping.qual.Bottom as bottom.
+ *
+ * @see TestReflectSibling1, TestReflectSibling2
+ */
+@SubtypeOf({TestReflectSibling1.class, TestReflectSibling2.class})
+@Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})
+@TargetLocations({TypeUseLocation.EXPLICIT_LOWER_BOUND, TypeUseLocation.EXPLICIT_UPPER_BOUND})
+@DefaultFor(TypeUseLocation.LOWER_BOUND)
+public @interface TestReflectBottom {}