+++ framework/src/test/java/org/checkerframework/framework/testchecker/reflection/qual/TestReflectSibling1.java
@@ -1,0 +1,16 @@
+package org.checkerframework.framework.testchecker.reflection.qual;
+
+import org.checkerframework.framework.qual.SubtypeOf;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Target;
+
+/**
+ * Toy type system for testing reflection resolution. Uses
+ * org.checkerframework.common.subtyping.qual.Bottom as bottom.
+ *
+ * @see TestReflectTop, TestReflectSibling2
+ */
+@SubtypeOf(TestReflectTop.class)
+@Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})
+public @interface TestReflectSibling1 {}