+++ framework/src/test/java/org/checkerframework/framework/testchecker/reflection/qual/TestReflectTop.java
@@ -1,0 +1,18 @@
+package org.checkerframework.framework.testchecker.reflection.qual;
+
+import org.checkerframework.framework.qual.DefaultQualifierInHierarchy;
+import org.checkerframework.framework.qual.SubtypeOf;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Target;
+
+/**
+ * Toy type system for testing reflection resolution. Uses
+ * org.checkerframework.common.subtyping.qual.Bottom as bottom.
+ *
+ * @see TestReflectSibling1, TestReflectSibling2
+ */
+@SubtypeOf({})
+@Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})
+@DefaultQualifierInHierarchy
+public @interface TestReflectTop {}